--- a/packages/babel-generator/benchmark/real-case/jquery.mjs
+++ b/packages/babel-generator/benchmark/real-case/jquery.mjs
@@ -15,7 +15,6 @@ function createInput(length) {
   );
 }
 
-current.default(createInput(1));
 function benchCases(name, implementation, options) {
   for (const length of [1, 2]) {
     const input = createInput(length);