… packages/babel-generator/benchmark/many-identifiers/1-length.mjs
… packages/babel-generator/benchmark/many-identifiers/25-length.mjs
… packages/babel-generator/benchmark/many-identifiers/bench.mjs
… packages/babel-generator/benchmark/real-case/jquery-3.6.txt
… packages/babel-generator/benchmark/util.mjs
… packages/babel-generator/package.json
… packages/babel-generator/src/node/parentheses.ts
… yarn.lock