<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | 
| Patch: Bug Fix?          |
| Major: Breaking Change?  |
| Minor: New Feature?      |
| Tests Added + Pass?      | Yes
| Documentation PR Link    | <!-- If only readme change, add `[skip ci]` to your commits -->
| Any Dependency Changes?  |
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->
This PR improves the `@babel/generator` performance. 

- The Buffer storage backend is switched to string, it was a micro-optimization in 6 yrs ago. But now the string approach (more intuitive) is faster according to the benchmark
- The popular path `Buffer#_endsWith` are rewritten to avoid scanning the `_queue`
- String comparison like `string[i] === "another one byte string"` is replaced by `string.charCodeAt(i) === charCodes`
- Access the "isQueries" in `@babel/types` is extremely slow. In this PR we hoist them to top level. Unfortunately we can not use the import namespace specifiers such as `import { isIdentifier } from "@babel/types"` because Babel replaces them to `types.isIdentifier`.

See commit messages for commit-by-commit performance reports.

In a real world case `jquery-3.6.txt`, we see 50% performance improvements compared to `@babel/generator` 7.14.5

```
baseline 1 jquery 3.6: 11.74 ops/sec ±12.2% (85ms)
baseline 2 jquery 3.6: 6.89 ops/sec ±5.4% (145ms)
current 1 jquery 3.6: 19.97 ops/sec ±9.56% (50ms)
current 2 jquery 3.6: 10.27 ops/sec ±4.38% (97ms)
```

<details>
  <summary>Many empty statements</summary>

```
baseline 256 empty statements: 3_455 ops/sec ±18.71% (0.289ms)
baseline 512 empty statements: 2_029 ops/sec ±1.29% (0.493ms)
baseline 1024 empty statements: 1_006 ops/sec ±1.37% (0.994ms)
baseline 2048 empty statements: 482 ops/sec ±2.01% (2.074ms)
current 256 empty statements: 5_334 ops/sec ±22.11% (0.187ms)
current 512 empty statements: 3_085 ops/sec ±1.73% (0.324ms)
current 1024 empty statements: 1_541 ops/sec ±1.73% (0.649ms)
current 2048 empty statements: 751 ops/sec ±1.95% (1.332ms)
```
</details>

<details>
  <summary>Many 1-length identifiers</summary>

```
baseline 256 1-length identifiers: 1_686 ops/sec ±30.51% (0.593ms)
baseline 512 1-length identifiers: 1_079 ops/sec ±1.38% (0.927ms)
baseline 1024 1-length identifiers: 533 ops/sec ±1.64% (1.876ms)
baseline 2048 1-length identifiers: 267 ops/sec ±1.67% (3.749ms)
current 256 1-length identifiers: 2_788 ops/sec ±25.92% (0.359ms)
current 512 1-length identifiers: 1_587 ops/sec ±1.34% (0.63ms)
current 1024 1-length identifiers: 792 ops/sec ±1.58% (1.262ms)
current 2048 1-length identifiers: 365 ops/sec ±3.4% (2.742ms)
```
</details>

<details>
  <summary>Many 25-length identifiers</summary>

```
baseline 256 25-length identifiers: 1_792 ops/sec ±25.73% (0.558ms)
baseline 512 25-length identifiers: 984 ops/sec ±2.46% (1.016ms)
baseline 1024 25-length identifiers: 520 ops/sec ±2.24% (1.924ms)
baseline 2048 25-length identifiers: 254 ops/sec ±1.91% (3.932ms)
current 256 25-length identifiers: 2_435 ops/sec ±49.59% (0.411ms)
current 512 25-length identifiers: 1_555 ops/sec ±1.78% (0.643ms)
current 1024 25-length identifiers: 751 ops/sec ±1.94% (1.331ms)
current 2048 25-length identifiers: 363 ops/sec ±1.65% (2.753ms)
```
</details>
<a href="https://gitpod.io/#https://github.com/babel/babel/pull/13593"><img src="https://gitpod.io/button/open-in-gitpod.svg"/></a>